+ simplify test_scenario

+++ crates/core/src/db/mock.rs
@@ -14,7 +14,7 @@ impl DbOps for MockDb {
         Ok(())
     }
 
-    fn insert_run(&self, _timestamp: u64, _tx_count: usize, _scenario_name: &str) -> Result<u64> {
+    fn insert_run(&self, _timestamp: u64, _tx_count: u64, _scenario_name: &str) -> Result<u64> {
         Ok(0)
     }
