+++ crates/core/src/generator/types.rs
@@ -114,5 +114,5 @@ pub enum PlanType<F: Fn(NamedTxRequest) -> CallbackResult> {
     /// Run setup steps, triggering a callback after each tx is sent.
     Setup(F),
     /// Spam with a number of txs and trigger a callback after each one is sent.
-    Spam(usize, F),
+    Spam(u64, F),
 }
